复盘修正第三波: 工程质量 + 技能升级落地
- CI: .github/workflows/ci.yml (build-test 零 warning 门禁 + UBSan/fuzz/TSan)
- fuzz_parsers: 随机畸形输入喂 rjson_stream/md/rbin (UBSan 20000 轮通过,
  抓到 rbin rd_u32 左移 int 溢出并修复)
- Makefile: make check (零 warning+全绿) / make tsan / make fuzz / make check-bench
- scripts/check_bench.py: 关键数字阈值回归 (rbin-load<50ms/SPSC>5Mops/高亮<2000ns/流式>100x/JSON>1.5x)
- docs/MEMORY.md: 三种内存来源所有权约定 (arena/owned_buf/mmap)
- docs/CHECKLIST.md: C 引擎提交前审查清单 (10 条踩坑易错点)

测试: 73/73 ALL PASSED, 零 warning, fuzz 20000 轮 UBSan 干净

+++ src/data/rbin.c
@@ -43,8 +43,8 @@ static uint8_t rd_u8(Rd *r) {
 }
 static uint32_t rd_u32(Rd *r) {
     if (r->off + 4 > r->len) { r->err = 1; return 0; }
-    uint32_t v = (uint32_t)(r->p[r->off] | (r->p[r->off + 1] << 8) |
-                            (r->p[r->off + 2] << 16) | (r->p[r->off + 3] << 24));
+    uint32_t v = (uint32_t)r->p[r->off] | ((uint32_t)r->p[r->off + 1] << 8) |
+                 ((uint32_t)r->p[r->off + 2] << 16) | ((uint32_t)r->p[r->off + 3] << 24);
     r->off += 4;
     return v;
 }
